Make floating geometry authoritative

+++ src/main/scala/com/serenity/ui/renderer/SurfaceContentResolver.scala
@@ -60,7 +60,7 @@ object SurfaceContentResolver:
     content: SurfaceContent,
     rect: LayoutRect,
     mode: SurfaceRenderMode,
-    itemGapRows: Int = 0
+    itemGapRows: Double = 0.0
   ): ResolvedSurfaceContent =
     content match
       case SurfaceContent.StartPage(_) =>
@@ -377,7 +377,7 @@ object SurfaceContentResolver:
     runner: com.serenity.command.CommandRunner,
     rect: LayoutRect,
     mode: SurfaceRenderMode,
-    itemGapRows: Int
+    itemGapRows: Double
   ): ResolvedSurfaceContent =
     if !runner.isActive then ResolvedSurfaceContent(titleFor(mode, "commands"))
     else
@@ -448,7 +448,7 @@ object SurfaceContentResolver:
     previewOnly: Boolean,
     rect: LayoutRect,
     mode: SurfaceRenderMode,
-    itemGapRows: Int
+    itemGapRows: Double
   ): ResolvedSurfaceContent =
     val group         = runner.submenuGroup(groupId)
     val submenuState  = runner.activeSubmenu.filter(_.groupId == groupId)
@@ -1001,7 +1001,7 @@ object SurfaceContentResolver:
     menu: ContextMenu,
     rect: LayoutRect,
     mode: SurfaceRenderMode,
-    itemGapRows: Int
+    itemGapRows: Double
   ): ResolvedSurfaceContent =
     val itemWindow = SurfaceFrameLayout(rect).itemWindow(
       itemCount = menu.items.size,
